requirements.txt
requests
types-requests
pychromecast
python-dotenv
google-auth
typing_extensions
SQLAlchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.3
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
aiohappyeyeballs >= 2.3.0
aiosignal >= 1.1.2
async-timeout >= 4.0, < 6.0 ; python_version < "3.11"
attrs >= 17.3.0
frozenlist >= 1.1.1
multidict >=4.5, < 7.0
propcache >= 0.2.0
yarl >= 1.17.0, < 2.0